+++ src/main/java/org/json/updated/parsers/JSONObject.java
@@ -1,0 +1,6 @@
+package org.json.updated.parsers;
+
+public interface JSONObject {
+    void handle();
+    String getName();
+}
