--- src/org/json/updated/parsers/JSONObject.java
@@ -1,6 +1,0 @@
-package org.json.updated.parsers;
-
-public interface JSONObject {
-    void handle();
-    String getName();
-}
